requirements.txt
python-telegram-bot==20.0
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
